- Fixed an error in part 1 (Was no alt for the title image)
An error still remains for the embedded video, but in the instructions, it was said to embed it as youtube gives it, so I assume I can safely ingnore the error. The error is about frameborder being obsolete and to use css instead, but nothing seems to be broken due to this.